.js파일을 .tsx파일로 변환
- 중첩되는 index.js파일을 삭제
- Board.js, Game.js, Square.js => Board.tsx, Game,tsx, Square.tsx
